refactor Deserialize implementation into content_serde module
 - use `btreemap!` in tests
 - move helper struct MessageDeHelper into content_serde
