+++ ruma-events/src/lib.rs
@@ -190,7 +190,7 @@ use serde::{
     de::{self, IgnoredAny},
     Deserialize, Serialize,
 };
-use serde_json::value::{Map as JsonObject, RawValue as RawJsonValue, Value as JsonValue};
+use serde_json::value::RawValue as RawJsonValue;
 
 use self::room::redaction::{RedactionEvent, SyncRedactionEvent};
 
@@ -518,18 +518,6 @@ pub struct EventDeHelper {
     /// If this `UnsignedData` contains a `redacted_because` key the event is
     /// immediately deserialized as a redacted event.
     pub unsigned: Option<UnsignedDeHelper>,
-}
-
-/// Helper struct to determine the msgtype from a `serde_json::value::RawValue`
-#[doc(hidden)]
-#[derive(Debug, Deserialize)]
-pub struct MessageDeHelper {
-    /// The message type field
-    pub msgtype: String,
-
-    /// Everything else in the json object
-    #[serde(flatten)]
-    pub remaining: JsonObject<String, JsonValue>,
 }
 
 /// Helper function for `serde_json::value::RawValue` deserialization.
